+++ CHANGELOG.md
@@ -12,8 +12,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 - The `get_terms` now have unique keys for the parameters to save memory.
-- `CombinatorialSpecification.expand_verified` will try to find a spec 
+- `CombinatorialSpecification.expand_verified` will try to find a spec
   allowing reverse rules if it fails without.
+- The labels assigned to nodes will always be in DFS order now.
 
 ### Fixed
 - Fixing bug creating non productive forest when expanding verified classes